pyproject.toml
[tool.poetry]
name = "wazzup"
version = "0.1.0"
description = ""
authors = ["Isaac Valdez <isaac_valdez@msn.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
pydantic = "^2.7.0"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
supabase = "^2.4.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
